Fixed LOTS of errors in both text.c's. 28 total

+++ include/text.h
@@ -14,7 +14,7 @@
 #define TF_UNSIGNED 128
 
 int m_printn(OUT char *str, IN int maxlen, IN int len, IN unsigned int n,
-		IN unsigned int base, IN int size, IN int flags, IN int precision);
+		IN unsigned int base, IN size_t size, IN int flags, IN int precision);
 int printf(IN const char *fmt, ...);
 int sprintf(OUT char *str, IN const char *fmt, ...);
 int snprintf(OUT char *str, IN size_t size, IN const char *fmt, ...);
